split generic single-file fixtures
